+++ Example/Example controllers/ReorderViewController.swift
@@ -21,15 +21,14 @@ class ReorderViewController: UIViewController, DTTableViewManageable, UITableVie
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        manager.startManaging(withDelegate: self)
         manager.register(StringCell.self)
         
         manager.memoryStorage.addItems(["Section 1 cell", "Section 1 cell"], toSection: 0)
         manager.memoryStorage.addItems(["Section 2 cell"], toSection: 1)
         manager.memoryStorage.addItems(["Section 3 cell", "Section 3 cell", "Section 3 cell"], toSection: 2)
         manager.canMove(StringCell.self, { _,_,_ in return true })
-        manager.editingStyle(for: StringCell.self, { _,_,_ in return .none })
-        
+        manager.editingStyle(forItem: String.self) { _,_ in .none }
+    
         manager.memoryStorage.setSectionHeaderModels(["Section 1", "Section 2", "Section 3"])
         manager.move(StringCell.self) { [weak manager] destination, _, _, source in
             manager?.memoryStorage.moveItemWithoutAnimation(from: source, to: destination)