+++ Example/Example controllers/Search/CoreDataSearchViewController.swift
@@ -30,11 +30,9 @@ class CoreDataSearchViewController: UIViewController, DTTableViewManageable {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        manager.startManaging(withDelegate: self)
-
         manager.register(BankCell.self)
         manager.storage = CoreDataStorage(fetchedResultsController: fetchResultsController)
-        (manager.tableViewUpdater as? TableViewUpdater)?.didUpdateContent = { [weak self] _ in
+        manager.tableViewUpdater?.didUpdateContent = { [weak self] _ in
             self?.tableView.isHidden = self?.tableView.numberOfSections == 0
         }