+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
-## [6.2.0](https://github.com/DenHeadless/DTTableViewManager/releases/tag/6.2.0)
+### Added
+
+* Anomaly detecting system for various errors in `DTTableViewManager`.
+
+### Changed
+
+* Calling `startManaging(withDelegate:_)` method is no longer required.
 
 ### Breaking
+
+* `viewFactoryErrorHandler` deprecated property on `DTTableViewManager` was removed. All previously reported errors and warnings are now a part of anomaly detecting system.
+
+## [6.2.0](https://github.com/DenHeadless/DTTableViewManager/releases/tag/6.2.0)
 
 * `editingStyle(for:_,_:)` method was replaced with `editingStyle(forItem:_,:_)` method, that accepts model and indexPath closure, without cell. Reason for that is that `UITableView` may call this method when cell is not actually on screen, in which case this event would not fire, and current editingStyle of the cell would be lost.